requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/requirements.in
#
asgiref==3.4.1
    # via django
attrs==21.4.0
    # via pytest
backports.zoneinfo==0.2.1
    # via django
django-extensions==3.1.5
    # via -r requirements/requirements.in
django-filter==21.1
    # via -r requirements/requirements.in
django==4.0.1
    # via
    #   -r requirements/requirements.in
    #   django-extensions
    #   django-filter
    #   djangorestframework
djangorestframework==3.13.1
    # via -r requirements/requirements.in
importlib-metadata==4.10.0
    # via markdown
iniconfig==1.1.1
    # via pytest
markdown==3.3.6
    # via -r requirements/requirements.in
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.3
    # via -r requirements/requirements.in
py==1.11.0
    # via pytest
pyparsing==3.0.6
    # via packaging
pytest-django==4.5.2
    # via -r requirements/requirements.in
pytest==6.2.5
    # via pytest-django
pytz==2021.3
    # via djangorestframework
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via pytest
zipp==3.7.0
    # via importlib-metadata
